+++ CHANGES.rst
@@ -3,6 +3,14 @@ Changes
 
 In next release ...
 
+- The `default` symbol in dynamic attributes is now
+  symbolic. Previously, it was assigned the string value of the
+  default attribute text. A similar change has been made for
+  switch/case expressions.
+
+- The built-in `attrs` dictionary of static element attributes now
+  correctly works with `tal:define`, etc.
+
 - Fix slice code generation compatibility issue on Python 3.9.
 
 3.8.0 (2020-06-25)