+++ CHANGES.rst
@@ -3,6 +3,8 @@ Changes
 
 In next release ...
 
+- Added code optimization to reduce sequential appends of static text.
+
 - The `default` symbol in dynamic attributes is now
   symbolic. Previously, it was assigned the string value of the
   default attribute text. A similar change has been made for